[Utility] Moved system message helper
