+++ Sources/Utility/Verbosity.swift
@@ -52,7 +52,6 @@ public class StandardErrorOutputStream: OutputStream {
 public var stderr = StandardErrorOutputStream()
 
 
-
 import func POSIX.prettyArguments
 
 private let ESC = "\u{001B}"
@@ -73,33 +72,6 @@ internal func printArgumentsIfVerbose(_ arguments: [String]) {
     }
 }
 
-
-import func libc.fflush
-import var libc.stdout
-import enum POSIX.Error
-
-public func system(_ arguments: String..., environment: [String:String] = [:], message: String?) throws {
-    var out = ""
-    do {
-        if Utility.verbosity == .Concise {
-            if let message = message {
-                print(message)
-                fflush(stdout)  // ensure we display `message` before git asks for credentials
-            }
-            try Utility.popen(arguments, redirectStandardError: true, environment: environment) { line in
-                out += line
-            }
-        } else {
-            try system(arguments, environment: environment)
-        }
-    } catch {
-        if verbosity == .Concise {
-            print(prettyArguments(arguments), to: &stderr)
-            print(out, to: &stderr)
-        }
-        throw error
-    }
-}
 
 internal func blue(_ input: String) -> String {
     return CSI + "34m" + input + CSI + "0m"
